core: move SQL value types and numeric into turso_core_types
Splits the Value half of core/types.rs (Value, ValueRef, Text, blobs,
FromValue, AggContext, SerialType) plus the numeric module into a new
turso_core_types crate, below the storage/vdbe/translate layers. The
cursor and record machinery (Cursor, ImmutableRecord, KeyInfo, record
comparators) stays in turso_core because it reaches into btree cursors,
sorters and collations; core/types.rs re-exports the moved items at their
old paths so nothing downstream changes.

The SQL operator semantics on Value (exec_add, exec_like, ...) stayed in
core/vdbe/value.rs, but inherent impls cannot live outside the type's
crate anymore, so they become the ValueExecOps extension trait. Callers
keep method syntax and only add a trait import.

Tests: cargo check -p turso_core (default, --tests, simulator,serde,json)
