+++ core/crates/types/numeric/mod.rs
@@ -1,4 +1,4 @@
-use crate::{types::AsValueRef, Value, ValueRef};
+use crate::{AsValueRef, Value, ValueRef};
 
 pub mod decimal;
 pub mod nonnan;
