+++ core/crates/types/numeric/nonnan.rs
